pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "strands-tools-community"
version = "0.1.0"
description = "Community tools for Strands Agent SDK: Deepgram, HubSpot, Microsoft Teams"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Eray Keskin", email = "eraykeskinmac@gmail.com"}
]
keywords = ["strands", "ai", "agents", "deepgram", "hubspot", "teams", "speech-to-text", "crm", "adaptive-cards"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "strands-agents>=0.1.0",
    "deepgram-sdk>=3.0.0",
    "requests>=2.31.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
demo = [
    "prompt-toolkit>=3.0.0",
    "halo>=0.0.31",
    "colorama>=0.4.6",
]

[project.urls]
Homepage = "https://github.com/eraykeskinmac/strands-tools-community"
Documentation = "https://github.com/eraykeskinmac/strands-tools-community"
Repository = "https://github.com/eraykeskinmac/strands-tools-community"
"Bug Reports" = "https://github.com/eraykeskinmac/strands-tools-community/issues"

[tool.setuptools]
packages = ["strands_tools_community", "strands_tools_community.templates"]

